requirements.txt
from_root
openpyxl
seaborn
plotly
streamlit==1.24.0
python-dotenv
numpy
pandas
flask
-e .
